pyproject.toml
[project]
name = "tower-of-babel-video"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "typer",
    "fastapi",
    "uvicorn",
    "alembic",
    "google-genai",
    "python-dotenv"
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "coverage",
    "faker",
    "hypothesis",
]
lint = [
    "mypy",
    "ruff",
]

[tool.uv]
default-groups = [
    "dev",
    "lint",
]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
strict = true

[tool.ruff]
target-version = "py313"

lint.select = [
    "A",    # flake8-builtins
    "ARG",  # flake8-unused-arguments
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "PT",   # flake8-pytest-style
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "SIM",  # flake8-simplify
    "T20",  # flake8-print
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.coverage.run]
branch = true
source = ["tests", "src/core", "src/infra", "src/runner"]

[tool.coverage.report]
skip_empty = true
skip_covered = true
show_missing = true
